Changelog For planned stable release of 3.4
Hello Jon and the Stencyl team! Its Donni11.
As we get closer to the stable release of 3.4 there are a lot of fixes going out in the newer beta releases. 
I've decided to create a changelog of all the new enhancement and fixes and fixes that are pending.

Forum link:  http://community.stencyl.com/index.php/topic,45126.0.html