Minor but important adjustment
I removed this :

 "Sure! Tell us in this forum topic."

I requested Stencyl to create a new topic for this : http://community.stencyl.com/index.php/topic,6386.msg263355.html#msg263355
because it was getting ridiculous. It was became a  "Recommend a behavior that you want someone (other than yourself) to create for you". So Stencyl locked the topic. Now this ^ doesn't apply anymore. So I replaced it by suggest a behavior in the suggestions issue tracker instead.