Rephrase parts of intro.md
I've never seen or heard anyone use "in a gist" before. I assume you meant something like "in short", "the basic points of", "cogent distinction", in which case "in essence" is appropriate in my book. In addition, "each" acts like a map function in English, referring to each individual element of a list at a time. "such as" could be replaced with "for example", so you should separate it from what it refers to. Finally, I deliberately placed the period at the end of the first bullet point in the summary in order to be logically consistent.