Remove changelog article and news section
The changelog here hasn't been updated in a long time, and it's not
a great place anyway for something that's updated with every build.

If a changelog is included here, it would have to be a more terse
summary of major releases instead of detailed build-by-build changes.

The news section never really took off and isn't visible on the site,
so remove it from here for now.